ContainerIO: Redefined ImageContainer_GetNumViews to return an unsigned 8 bit int instead of 64 bit one, cuz there's not gonna more than a few views.

+++ Library/include/ContainerIO.h
@@ -317,7 +317,7 @@ extern "C" {
      @abstract                             "Gets the number of views in an ImageContainer".
      @param                 Image          "A pointer to the instance of an ImageContainer in question".
      */
-    uint64_t                ImageContainer_GetNumViews(ImageContainer *Image);
+    uint8_t                 ImageContainer_GetNumViews(ImageContainer *Image);
     
     /*!
      @abstract                             "Gets the views index for the specified mask".
